pyproject.toml
[project]
name = "drf-vps-api"
version = "0.1.0"
description = ""
authors = [
    {name = "kot9pa",email = "kot9paa@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "djangorestframework (>=3.15.2,<4.0.0)",
    "django (>=5.1.5,<6.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "dj-database-url (>=2.3.0,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-extensions (>=3.2.3,<4.0.0)",
    "django-cors-headers (>=4.6.0,<5.0.0)",
    "django-filter (>=24.3,<25.0)",
    "drf-yasg (>=1.21.8,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
